[#2] We're getting close to showing model
However, even though loading and transferring seems to work -we even added a rotating camera- the model isn't shown yet. Maybe the next step would be to add a basic movement system for the camera, s.t. we can freely look around and find the model in case we spawn inside and the culling gets us.