Working on yet _another_ Pipeline implementation.
This way, though, we made sure to neatly separate Shaders in a ShaderPool. Will have to see if this idea of a separate Constructors is the way to goooooooooo wait a minute... What if we discard the idea of a PipelinePool altogether? Instead, we make a small wrapper around the cache, and then use a PipelineConstruct to spawn pipelines - each time, the constructor can be subtely modified to keep the pipelines the same. The cache can then take over the similar bits.